Reverting changes, as requested.
